+++ src/main/kotlin/io/github/woolph/gradle/QualityCheckExtension.kt
@@ -1,4 +1,6 @@
 /* Copyright 2023 ENGEL Austria GmbH */
 package io.github.woolph.gradle
 
-@Suppress("UnnecessaryAbstractClass") abstract class QualityCheckExtension
+import org.gradle.api.plugins.ExtensionAware
+
+@Suppress("UnnecessaryAbstractClass") abstract class QualityCheckExtension: ExtensionAware