+++ src/main/kotlin/io/github/woolph/gradle/QualityCheckPlugin.kt
@@ -7,8 +7,8 @@ import io.github.woolph.gradle.licensecheck.LicenseCheckExtension
 import io.github.woolph.gradle.licensecheck.LicenseCheckExtension.Companion.applyLicenseCheckExtension
 import org.gradle.api.Plugin
 import org.gradle.api.Project
-import org.gradle.api.plugins.ExtensionAware
 import org.gradle.kotlin.dsl.create
+import org.gradle.kotlin.dsl.named
 
 @Suppress("unused")
 class QualityCheckPlugin : Plugin<Project> {
@@ -16,11 +16,22 @@ class QualityCheckPlugin : Plugin<Project> {
     private lateinit var dependencyCheckExtension: DependencyCheckExtension
     private lateinit var licenseCheckExtension: LicenseCheckExtension
 
-    override fun apply(project: Project): Unit =
+    override fun apply(project: Project) {
         project.run {
-            qualityCheckExtension = extensions.create("qualityCheck", QualityCheckExtension::class)
+            plugins.apply("jacoco")
 
-            applyDependencyCheckExtension(qualityCheckExtension as ExtensionAware)
-            applyLicenseCheckExtension(qualityCheckExtension as ExtensionAware)
+            val jacocoTestReport = tasks.named<org.gradle.testing.jacoco.tasks.JacocoReport>("jacocoTestReport") {
+                reports {
+                    xml.required.set(true)
+                }
+            }
+            tasks.named<org.gradle.api.tasks.testing.Test>("test") {
+                finalizedBy(jacocoTestReport)
+            }
+
+            qualityCheckExtension = extensions.create("qualityCheck", QualityCheckExtension::class)
+            applyDependencyCheckExtension(qualityCheckExtension)
+            applyLicenseCheckExtension(qualityCheckExtension)
         }
+    }
 }